# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
    ObservabilityQueryResponse:
      type: object
      description: "PromQL 代理查询结果；不暴露底层 Prometheus 地址。"
      required: [query, result_type, results, dev_profile]
      properties:
        query:       { type: string }
        result_type: { type: string, enum: [vector, matrix, scalar, string] }
        results:
          type: array
          items:
            type: object
            required: [metric, value]
            properties:
              metric:
                type: object
                additionalProperties: { type: string }
              value: { type: number, format: double }
              timestamp: { type: string, format: date-time, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
    ObservabilityAlertRule:
      type: object
      required: [id, tenant_id, name, promql, severity, state, created_at, updated_at]
      properties:
        id:          { type: string }
        tenant_id:   { type: string }
        name:        { type: string }
        promql:      { type: string }
        duration:    { type: string, default: 5m }
        severity:    { type: string, enum: [info, warning, critical] }
        labels:
          type: object
          additionalProperties: { type: string }
        annotations:
          type: object
          additionalProperties: { type: string }
        enabled:     { type: boolean, default: true }
        state:       { type: string, enum: [active, disabled, deleted] }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
    ObservabilityAlertRuleListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ObservabilityAlertRule' }
        total:       { type: integer }
        next_cursor: { type: string, nullable: true }
    CreateObservabilityAlertRuleRequest:
      type: object
      required: [idempotency_key, name, promql, severity]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128 }
        name:           { type: string, minLength: 1, maxLength: 128 }
        promql:         { type: string, minLength: 1 }
        duration:       { type: string, default: 5m }
        severity:       { type: string, enum: [info, warning, critical] }
        labels:
          type: object
          additionalProperties: { type: string }
        annotations:
          type: object
          additionalProperties: { type: string }
        enabled:        { type: boolean, default: true }
    UpdateObservabilityAlertRuleRequest:
      type: object
      required: [idempotency_key]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128 }
        name:           { type: string, minLength: 1, maxLength: 128 }
        promql:         { type: string, minLength: 1 }
        duration:       { type: string }
        severity:       { type: string, enum: [info, warning, critical] }
        labels:
          type: object
          additionalProperties: { type: string }
        annotations:
          type: object
          additionalProperties: { type: string }
        enabled:        { type: boolean }
...
    ReportTokenUsageRequest:
      type: object
      required: [idempotency_key, source, model, input_tokens, output_tokens]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128 }
        source:          { type: string, description: "上报来源，如 model-service、inference-gateway" }
        model:           { type: string }
        input_tokens:    { type: integer, format: int64, minimum: 0 }
        output_tokens:   { type: integer, format: int64, minimum: 0 }
        request_id:      { type: string, nullable: true }
        instance_id:     { type: string, nullable: true }
        occurred_at:     { type: string, format: date-time, nullable: true }
        labels:
          type: object
          additionalProperties: { type: string }
    TokenUsageReport:
      type: object
      required: [id, tenant_id, source, model, input_tokens, output_tokens, total_tokens, state, created_at]
      properties:
        id:            { type: string }
        tenant_id:     { type: string }
        source:        { type: string }
        model:         { type: string }
        input_tokens:  { type: integer, format: int64 }
        output_tokens: { type: integer, format: int64 }
        total_tokens:  { type: integer, format: int64 }
        request_id:    { type: string, nullable: true }
        instance_id:   { type: string, nullable: true }
        state:         { type: string, enum: [accepted, duplicate] }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at:    { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: 未认证或 Token 无效（code=UNAUTHORIZED）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权访问该资源（code=FORBIDDEN）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: 请求参数不合法（code=BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /observability/query:
    get:
      operationId: queryObservability
      summary: PromQL 代理查询
      description: 通过 Core 代理 PromQL 查询，不暴露底层 Prometheus 地址。
      tags: [Observability]
      x-ani-rbac-scope: "scope:observability:read"
      parameters:
        - { name: query, in: query, required: true, schema: { type: string, minLength: 1 } }
        - { name: time, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: timeout, in: query, required: false, schema: { type: string, default: 30s } }
      responses:
        "200":
          description: PromQL 查询结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservabilityQueryResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /observability/alert-rules:
    get:
      operationId: listObservabilityAlertRules
      summary: 列出告警规则
      tags: [Observability]
      x-ani-rbac-scope: "scope:observability:read"
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 告警规则列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservabilityAlertRuleListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createObservabilityAlertRule
      summary: 创建告警规则
      tags: [Observability]
      x-ani-rbac-scope: "scope:observability:write"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateObservabilityAlertRuleRequest' }
      responses:
        "201":
          description: 告警规则已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservabilityAlertRule' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /observability/alert-rules/{rule_id}:
    get:
      operationId: getObservabilityAlertRule
      summary: 获取告警规则
      tags: [Observability]
      x-ani-rbac-scope: "scope:observability:read"
      parameters:
        - { name: rule_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 告警规则详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservabilityAlertRule' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateObservabilityAlertRule
      summary: 更新告警规则
      tags: [Observability]
      x-ani-rbac-scope: "scope:observability:write"
      parameters:
        - { name: rule_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateObservabilityAlertRuleRequest' }
      responses:
        "200":
          description: 告警规则已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservabilityAlertRule' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteObservabilityAlertRule
      summary: 删除告警规则
      tags: [Observability]
      x-ani-rbac-scope: "scope:observability:write"
      parameters:
        - { name: rule_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 告警规则已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservabilityAlertRule' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── 计量 ────────────────────────────────────────────────────────────────────
  /metering/usage:
    get:
      summary: 查询租户用量统计
      tags: [Metering]
      parameters:
        - { name: start_time, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: end_time,   in: query, required: true, schema: { type: string, format: date-time } }
        - { name: resource_type, in: query, schema: { type: string } }
        - { name: group_by, in: query, schema: { type: string, enum: [resource_type, az, day, hour] } }
      responses:
        "200":
          description: 租户用量统计
          content:
            application/json:
              schema:
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        resource_type: { type: string }
                        total_quantity: { type: number }
                        unit:           { type: string }
                        period:         { type: string }
                  total: { type: integer }
                  dev_profile:
                    $ref: '#/components/schemas/CoreDevProfileInfo'
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /metering/token-usage:
    post:
      operationId: reportTokenUsage
      summary: 上报 Token 用量
      description: Services 层通过 Core API 上报 Token 计量事件；客户端重试必须复用同一个 idempotency_key。
      tags: [Metering]
      x-ani-rbac-scope: "scope:metering:write"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportTokenUsageRequest' }
      responses:
        "202":
          description: Token 用量事件已接受
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenUsageReport' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
...